pyproject.toml
[project]
name = "image-optimization-service"
version = "0.1.0"
description = ""
authors = [
    {name = "BlackDChase",email = "blackdragonchase@gmail.com"}
]
readme = "README.md"
requires-python = "3.13.5"
dependencies = [
    "django (==5.2.3)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-prometheus (>=2.4.0,<3.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pylint (>=3.3.7,<4.0.0)",
    "django-extensions (>=4.1,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
